WIP plumbing LogDb
