+++ crates/re_arrow_store/src/store_write.rs
@@ -1,21 +1,14 @@
 use std::collections::HashMap;
 use std::sync::Arc;
 
-use anyhow::{anyhow, ensure};
-use arrow2::array::{
-    new_empty_array, Array, Int64Vec, ListArray, MutableArray, StructArray, UInt64Vec,
-};
+use arrow2::array::{new_empty_array, Array, Int64Vec, ListArray, MutableArray, UInt64Vec};
 use arrow2::buffer::Buffer;
-use arrow2::chunk::Chunk;
 use arrow2::compute::concatenate::concatenate;
-use arrow2::datatypes::{DataType, Schema};
-use polars::prelude::IndexOfSchema;
+use arrow2::datatypes::DataType;
 
 use re_log::debug;
-use re_log_types::external::arrow2_convert::deserialize::arrow_array_deserialize_iterator;
 use re_log_types::{
     ComponentNameRef, ObjPath as EntityPath, TimeInt, TimePoint, TimeRange, Timeline,
-    ENTITY_PATH_KEY,
 };
 
 use crate::{
@@ -31,18 +24,14 @@ impl DataStore {
     /// - the entity path,
     /// - the targeted timelines & timepoints,
     /// - and all the components data.
-    pub fn insert(&mut self, schema: &Schema, msg: &Chunk<Box<dyn Array>>) -> anyhow::Result<()> {
-        let ent_path = schema
-            .metadata
-            .get(ENTITY_PATH_KEY)
-            .ok_or_else(|| anyhow!("expect entity path in top-level message's metadata"))
-            .map(|path| EntityPath::from(path.as_str()))?;
-        let ent_path_hash = *ent_path.hash();
-
-        let timelines = extract_timelines(schema, msg)?;
-        let components = extract_components(schema, msg)?;
-
+    pub fn insert(
+        &mut self,
+        ent_path: &EntityPath,
+        time_point: &TimePoint,
+        components: impl ExactSizeIterator<Item = (ComponentNameRef<'static>, Box<dyn Array>)>,
+    ) -> anyhow::Result<()> {
         // TODO(cmc): sort the "instances" component, and everything else accordingly!
+        let ent_path_hash = *ent_path.hash();
 
         let mut indices = HashMap::with_capacity(components.len());
         for (name, component) in components {
@@ -50,11 +39,11 @@ impl DataStore {
                 ComponentTable::new(name.to_owned(), component.data_type().clone())
             });
 
-            let row_idx = table.insert(&self.config, &timelines, component)?;
+            let row_idx = table.insert(&self.config, time_point.iter(), component.as_ref())?;
             indices.insert(name, row_idx);
         }
 
-        for (timeline, time) in &timelines {
+        for (timeline, time) in time_point.iter() {
             let ent_path = ent_path.clone(); // shallow
             let index = self
                 .indices
@@ -65,51 +54,6 @@ impl DataStore {
 
         Ok(())
     }
-}
-
-fn extract_timelines(
-    schema: &Schema,
-    msg: &Chunk<Box<dyn Array>>,
-) -> anyhow::Result<Vec<(Timeline, TimeInt)>> {
-    let timelines = schema
-        .index_of("timelines") // TODO(cmc): maybe at least a constant or something
-        .and_then(|idx| msg.columns().get(idx))
-        .ok_or_else(|| anyhow!("expect top-level `timelines` field`"))?;
-
-    let mut timepoints_iter = arrow_array_deserialize_iterator::<TimePoint>(timelines.as_ref())?;
-
-    let timepoint = timepoints_iter
-        .next()
-        .ok_or_else(|| anyhow!("No rows in timelines."))?;
-
-    ensure!(
-        timepoints_iter.next().is_none(),
-        "Expected a single TimePoint, but found more!"
-    );
-
-    Ok(timepoint.into_iter().collect())
-}
-
-fn extract_components<'data>(
-    schema: &Schema,
-    msg: &'data Chunk<Box<dyn Array>>,
-) -> anyhow::Result<Vec<(ComponentNameRef<'data>, &'data dyn Array)>> {
-    let components = schema
-        .index_of("components") // TODO(cmc): maybe at least a constant or something
-        .and_then(|idx| msg.columns().get(idx))
-        .ok_or_else(|| anyhow!("expect top-level `components` field`"))?;
-
-    let components = components
-        .as_any()
-        .downcast_ref::<StructArray>()
-        .ok_or_else(|| anyhow!("expect component values to be `StructArray`s"))?;
-
-    Ok(components
-        .fields()
-        .iter()
-        .zip(components.values())
-        .map(|(field, comp)| (field.name.as_str(), comp.as_ref()))
-        .collect())
 }
 
 // --- Indices ---
@@ -205,10 +149,10 @@ impl ComponentTable {
         }
     }
 
-    pub fn insert(
+    pub fn insert<'a>(
         &mut self,
         config: &DataStoreConfig,
-        timelines: &[(Timeline, TimeInt)],
+        timelines: impl IntoIterator<Item = (&'a Timeline, &'a TimeInt)>,
         data: &dyn Array,
     ) -> anyhow::Result<RowIndex> {
         // All component tables spawn with an initial bucket at row offset 0, thus this cannot
@@ -280,10 +224,9 @@ impl ComponentBucket {
             data,
         }
     }
-
-    pub fn insert(
+    pub fn insert<'a>(
         &mut self,
-        timelines: &[(Timeline, TimeInt)],
+        timelines: impl IntoIterator<Item = (&'a Timeline, &'a TimeInt)>,
         data: &dyn Array,
     ) -> anyhow::Result<RowIndex> {
         for (timeline, time) in timelines {
