+++ crates/re_log_types/Cargo.toml
@@ -8,7 +8,7 @@ publish = false
 
 
 [features]
-default = ["arrow_datagen"]
+default = ["arrow_datagen", "anyhow"]
 
 ## Enable loading data from a file.
 load = ["anyhow", "rmp-serde", "serde", "zstd", "ruzstd"]
