+++ crates/re_log_types/src/arrow_msg.rs
@@ -3,9 +3,18 @@
 //! We have custom implementations of [`serde::Serialize`] and [`serde::Deserialize`] that wraps
 //! the inner Arrow serialization of [`Schema`] and [`Chunk`].
 
-use arrow2::{array::Array, chunk::Chunk, datatypes::Schema};
+use std::collections::BTreeMap;
 
-use crate::MsgId;
+use arrow2::{
+    array::{Array, StructArray},
+    chunk::Chunk,
+    datatypes::{DataType, Field, Schema},
+};
+use arrow2_convert::{field::ArrowField, serialize::TryIntoArrow};
+
+use crate::{ComponentNameRef, MsgId, ObjPath, TimeInt, TimePoint, Timeline, ENTITY_PATH_KEY};
+
+use anyhow::{anyhow, ensure};
 
 /// Message containing an Arrow payload
 #[must_use]
@@ -99,14 +108,141 @@ impl<'de> serde::Deserialize<'de> for ArrowMsg {
     }
 }
 
+// ----------------------------------------------------------------------------
+
+pub fn pack_components(
+    components: impl Iterator<Item = (Schema, Box<dyn Array>)>,
+) -> (Schema, StructArray) {
+    let (component_schemas, component_cols): (Vec<_>, Vec<_>) = components.unzip();
+    let component_fields = component_schemas
+        .into_iter()
+        .flat_map(|schema| schema.fields)
+        .collect();
+
+    let packed = StructArray::new(DataType::Struct(component_fields), component_cols, None);
+
+    let schema = Schema {
+        fields: [Field::new("components", packed.data_type().clone(), false)].to_vec(),
+        ..Default::default()
+    };
+
+    (schema, packed)
+}
+
+/// Build a single log message
+pub fn build_message(
+    ent_path: &ObjPath,
+    timepoint: &TimePoint,
+    components: impl IntoIterator<Item = (ComponentNameRef<'static>, Schema, Box<dyn Array>)>,
+) -> (Schema, Chunk<Box<dyn Array>>) {
+    let mut schema = Schema::default();
+    let mut cols: Vec<Box<dyn Array>> = Vec::new();
+
+    schema.metadata = BTreeMap::from([(ENTITY_PATH_KEY.into(), ent_path.to_string())]);
+
+    // Build & pack timelines
+    let timelines_field = Field::new("timelines", TimePoint::data_type(), false);
+    let timelines_col = [timepoint].try_into_arrow().unwrap();
+
+    schema.fields.push(timelines_field);
+    cols.push(timelines_col);
+
+    // Build & pack components
+    let (components_schema, components_data) = pack_components(
+        components
+            .into_iter()
+            .map(|(_, schema, data)| (schema, data)),
+    );
+    schema.fields.extend(components_schema.fields);
+    schema.metadata.extend(components_schema.metadata);
+    cols.push(components_data.boxed());
+
+    (schema, Chunk::new(cols))
+}
+
+pub fn extract_message<'data>(
+    schema: &'_ Schema,
+    msg: &'data Chunk<Box<dyn Array>>,
+) -> anyhow::Result<(
+    ObjPath,
+    TimePoint,
+    Vec<(ComponentNameRef<'data>, &'data dyn Array)>,
+)> {
+    let ent_path = schema
+        .metadata
+        .get(ENTITY_PATH_KEY)
+        .ok_or_else(|| anyhow!("expect entity path in top-level message's metadata"))
+        .map(|path| ObjPath::from(path.as_str()))?;
+
+    let time_point = extract_timelines(schema, msg)?;
+    let components = extract_components(schema, msg)?;
+
+    Ok((ent_path, time_point, components))
+}
+
+/// Extract a [`TimePoint`] from the "timelines" column
+pub fn extract_timelines(
+    schema: &Schema,
+    msg: &Chunk<Box<dyn Array>>,
+) -> anyhow::Result<TimePoint> {
+    use arrow2_convert::deserialize::arrow_array_deserialize_iterator;
+
+    let timelines = schema
+        .fields
+        .iter()
+        .position(|f| f.name == "timelines") // TODO(cmc): maybe at least a constant or something
+        .and_then(|idx| msg.columns().get(idx))
+        .ok_or_else(|| anyhow!("expect top-level `timelines` field`"))?;
+
+    let mut timepoints_iter = arrow_array_deserialize_iterator::<TimePoint>(timelines.as_ref())?;
+
+    let timepoint = timepoints_iter
+        .next()
+        .ok_or_else(|| anyhow!("No rows in timelines."))?;
+
+    ensure!(
+        timepoints_iter.next().is_none(),
+        "Expected a single TimePoint, but found more!"
+    );
+
+    Ok(timepoint)
+}
+
+/// Extract the components from the message
+pub fn extract_components<'data>(
+    schema: &Schema,
+    msg: &'data Chunk<Box<dyn Array>>,
+) -> anyhow::Result<Vec<(ComponentNameRef<'data>, &'data dyn Array)>> {
+    let components = schema
+        .fields
+        .iter()
+        .position(|f| f.name == "components") // TODO(cmc): maybe at least a constant or something
+        .and_then(|idx| msg.columns().get(idx))
+        .ok_or_else(|| anyhow!("expect top-level `components` field`"))?;
+
+    let components = components
+        .as_any()
+        .downcast_ref::<StructArray>()
+        .ok_or_else(|| anyhow!("expect component values to be `StructArray`s"))?;
+
+    Ok(components
+        .fields()
+        .iter()
+        .zip(components.values())
+        .map(|(field, comp)| (field.name.as_str(), comp.as_ref()))
+        .collect())
+}
+
+// ----------------------------------------------------------------------------
+
 #[cfg(test)]
 #[cfg(feature = "serde")]
 mod tests {
     use serde_test::{assert_tokens, Token};
 
-    use super::{ArrowMsg, Chunk, MsgId, Schema};
+    use super::{build_message, ArrowMsg, Chunk, MsgId, Schema};
     use crate::{
-        datagen::{build_frame_nr, build_message, build_positions, build_rects},
+        datagen::{build_frame_nr, build_positions, build_rects},
         ObjPath, TimePoint,
     };
