+++ crates/re_log_types/src/datagen.rs
@@ -188,53 +188,3 @@ pub fn build_positions(nb_instances: usize) -> (ComponentNameRef<'static>, Schem
 
     ("positions", schema, data.boxed())
 }
-
-pub fn pack_components(
-    components: impl Iterator<Item = (Schema, Box<dyn Array>)>,
-) -> (Schema, StructArray) {
-    let (component_schemas, component_cols): (Vec<_>, Vec<_>) = components.unzip();
-    let component_fields = component_schemas
-        .into_iter()
-        .flat_map(|schema| schema.fields)
-        .collect();
-
-    let packed = StructArray::new(DataType::Struct(component_fields), component_cols, None);
-
-    let schema = Schema {
-        fields: [Field::new("components", packed.data_type().clone(), false)].to_vec(),
-        ..Default::default()
-    };
-
-    (schema, packed)
-}
-
-/// Build a single log message
-pub fn build_message(
-    ent_path: &EntityPath,
-    timepoint: &TimePoint,
-    components: impl IntoIterator<Item = (ComponentNameRef<'static>, Schema, Box<dyn Array>)>,
-) -> (Schema, Chunk<Box<dyn Array>>) {
-    let mut schema = Schema::default();
-    let mut cols: Vec<Box<dyn Array>> = Vec::new();
-
-    schema.metadata = BTreeMap::from([(ENTITY_PATH_KEY.into(), ent_path.to_string())]);
-
-    // Build & pack timelines
-    let timelines_field = Field::new("timelines", TimePoint::data_type(), false);
-    let timelines_col = [timepoint].try_into_arrow().unwrap();
-
-    schema.fields.push(timelines_field);
-    cols.push(timelines_col);
-
-    // Build & pack components
-    let (components_schema, components_data) = pack_components(
-        components
-            .into_iter()
-            .map(|(_, schema, data)| (schema, data)),
-    );
-    schema.fields.extend(components_schema.fields);
-    schema.metadata.extend(components_schema.metadata);
-    cols.push(components_data.boxed());
-
-    (schema, Chunk::new(cols))
-}
